feat(@bautajs/core,@bautajs/express,@bautajs/datasource-rest): logger (#142)
feat: Allows a logger to be passed as parameter to bauta. This is in the issue #138. 
feat: If no logger is passed, default logger is used. Note that in this PR tests continue testing only default case with 
feat: Simplifies logger model and eliminates unneded internal data structures (ContextLogger, BuilderLogger, references to moduleLogger, etc...)
fix: Refactored a lot to stop using a fixed static logger and instead use the unique bauta logger.
fix: Error message added in the error shown in the logs when a provider in bauta generates an error.
BREAKING CHANGE: Events are no longer emitted since they were not used.