fix(@bautajs/core): response validation (#235)
Response validation should validate the statusCodes of the schema.

If statusCode not found on the schema and there is no default response throw 500 error-.
If statusCode is found on the schema validate against that response schema.
If statusCode is not found on the schema but there is default validate against that response schema.